refactor `Open in Database Tools` of MySQL/PostgreSQL/Sql Server
