+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-database/build.gradle
@@ -1,4 +1,5 @@
 intellij {
+    plugins = ['com.intellij.database']
 }
 
 dependencies {
