+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-database/src/main/resources/META-INF/azure-intellij-plugin-database.xml
@@ -1,4 +1,5 @@
 <idea-plugin>
+    <depends optional="true" config-file="azure-intellij-plugin-database-dbtools.xml">com.intellij.database</depends>
     <extensions defaultExtensionNs="com.microsoft.tooling.msservices.intellij.azure">
         <connectorResourceType implementation="com.microsoft.azure.toolkit.intellij.database.mysql.connection.MySqlDatabaseResourceDefinition"/>
         <connectorResourceType implementation="com.microsoft.azure.toolkit.intellij.database.postgre.connection.PostgreSqlDatabaseResourceDefinition"/>
